% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\docType{class}
\name{Algorithm}
\alias{Algorithm}
\title{Algorithm}
\format{
An \code{R6Class} generator object
}
\description{
Algorithm Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{name}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Algorithm$new()}}
\item \href{#method-toJSON}{\code{Algorithm$toJSON()}}
\item \href{#method-fromJSON}{\code{Algorithm$fromJSON()}}
\item \href{#method-toJSONString}{\code{Algorithm$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Algorithm$fromJSONString()}}
\item \href{#method-clone}{\code{Algorithm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of Algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$new(description = NULL, id = NULL, name = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{(character) Description of the algorithm.}

\item{\code{id}}{(numeric) Unique integer identifying the algorithm.}

\item{\code{name}}{(character) Name of the algorithm.}

\item{\code{...}}{(list) local.optional.var}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert an instance of Algorithm to an R list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Convert an R JSON object to an R list type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$fromJSON(AlgorithmJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AlgorithmJson}}{(character) A JSON string representing an Algorithm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Convert an instance of Algorithm to a JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Create an instance of Algorithm from a JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$fromJSONString(AlgorithmJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AlgorithmJson}}{(character) A JSON string representing an Algorithm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Algorithm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
