% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_launch_spec.R
\docType{class}
\name{AlgorithmLaunchSpec}
\alias{AlgorithmLaunchSpec}
\title{AlgorithmLaunchSpec}
\format{
An \code{R6Class} generator object
}
\description{
AlgorithmLaunchSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{algorithm_name}}{character}

\item{\code{media_ids}}{list( integer ) \link{optional}}

\item{\code{media_query}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AlgorithmLaunchSpec$new()}}
\item \href{#method-toJSON}{\code{AlgorithmLaunchSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{AlgorithmLaunchSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{AlgorithmLaunchSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AlgorithmLaunchSpec$fromJSONString()}}
\item \href{#method-clone}{\code{AlgorithmLaunchSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of AlgorithmLaunchSpec
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgorithmLaunchSpec$new(
  algorithm_name,
  media_ids = NULL,
  media_query = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{algorithm_name}}{(character) Name of the algorithm to execute.}

\item{\code{media_ids}}{(vector) List of media IDs. Must supply media_query or media_ids.}

\item{\code{media_query}}{(character) Query string used to filter media IDs. If supplied, media_ids will be ignored.}

\item{\code{...}}{(list) local.optional.var}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert an instance of AlgorithmLaunchSpec to an R list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgorithmLaunchSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Convert an R JSON object to an R list type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgorithmLaunchSpec$fromJSON(AlgorithmLaunchSpecJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AlgorithmLaunchSpecJson}}{(character) A JSON string representing an AlgorithmLaunchSpec}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Convert an instance of AlgorithmLaunchSpec to a JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgorithmLaunchSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Create an instance of AlgorithmLaunchSpec from a JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgorithmLaunchSpec$fromJSONString(AlgorithmLaunchSpecJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AlgorithmLaunchSpecJson}}{(character) A JSON string representing an AlgorithmLaunchSpec}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlgorithmLaunchSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
