% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_spec.R
\docType{class}
\name{AnalysisSpec}
\alias{AnalysisSpec}
\title{AnalysisSpec}
\format{
An \code{R6Class} generator object
}
\description{
AnalysisSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_query}}{character}

\item{\code{name}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnalysisSpec$new()}}
\item \href{#method-toJSON}{\code{AnalysisSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{AnalysisSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{AnalysisSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AnalysisSpec$fromJSONString()}}
\item \href{#method-clone}{\code{AnalysisSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of AnalysisSpec
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSpec$new(data_query, name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_query}}{(character) Lucene query string used to retrieve entities to analyze.}

\item{\code{name}}{(character) Name of analysis.}

\item{\code{...}}{(list) local.optional.var}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Convert an instance of AnalysisSpec to an R list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Convert an R JSON object to an R list type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSpec$fromJSON(AnalysisSpecJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnalysisSpecJson}}{(character) A JSON string representing an AnalysisSpec}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Convert an instance of AnalysisSpec to a JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Create an instance of AnalysisSpec from a JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSpec$fromJSONString(AnalysisSpecJson)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AnalysisSpecJson}}{(character) A JSON string representing an AnalysisSpec}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
