% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_type.R
\docType{class}
\name{AttributeType}
\alias{AttributeType}
\title{AttributeType}
\format{
An \code{R6Class} generator object
}
\description{
AttributeType Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{autocomplete}}{\link{AutocompleteService} \link{optional}}

\item{\code{choices}}{list( character ) \link{optional}}

\item{\code{default}}{\link{AnyType} \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{dtype}}{character \link{optional}}

\item{\code{labels}}{list( character ) \link{optional}}

\item{\code{maximum}}{numeric \link{optional}}

\item{\code{minimum}}{numeric \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{order}}{integer \link{optional}}

\item{\code{required}}{character \link{optional}}

\item{\code{use_current}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AttributeType$new()}}
\item \href{#method-toJSON}{\code{AttributeType$toJSON()}}
\item \href{#method-fromJSON}{\code{AttributeType$fromJSON()}}
\item \href{#method-toJSONString}{\code{AttributeType$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AttributeType$fromJSONString()}}
\item \href{#method-clone}{\code{AttributeType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeType$new(
  autocomplete = NULL,
  choices = NULL,
  default = NULL,
  description = "",
  dtype = NULL,
  labels = NULL,
  maximum = NULL,
  minimum = NULL,
  name = NULL,
  order = 0,
  required = FALSE,
  use_current = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeType$fromJSON(AttributeTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeType$fromJSONString(AttributeTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
