% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_api}
\alias{get_api}
\title{Retrieves a \code{TatorAPI} instance using the given host and token.
@param host URL of host. Default is https://www.tatorapp.com.
@param token API token.
@return TatorApi object}
\usage{
get_api(host = "https://www.tatorapp.com", token = Sys.getenv("TATOR_TOKEN"))
}
\description{
Retrieves a \code{TatorAPI} instance using the given host and token.
@param host URL of host. Default is https://www.tatorapp.com.
@param token API token.
@return TatorApi object
}
