% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\docType{class}
\name{Localization}
\alias{Localization}
\title{Localization}
\format{
An \code{R6Class} generator object
}
\description{
Localization Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attributes}}{named list( \link{AnyType} ) \link{optional}}

\item{\code{email}}{character \link{optional}}

\item{\code{frame}}{integer \link{optional}}

\item{\code{height}}{numeric \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{media}}{integer \link{optional}}

\item{\code{meta}}{integer \link{optional}}

\item{\code{modified}}{character \link{optional}}

\item{\code{parent}}{numeric \link{optional}}

\item{\code{project}}{integer \link{optional}}

\item{\code{thumbnail_image}}{character \link{optional}}

\item{\code{u}}{numeric \link{optional}}

\item{\code{v}}{numeric \link{optional}}

\item{\code{version}}{integer \link{optional}}

\item{\code{width}}{numeric \link{optional}}

\item{\code{x}}{numeric \link{optional}}

\item{\code{y}}{numeric \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Localization$new()}}
\item \href{#method-toJSON}{\code{Localization$toJSON()}}
\item \href{#method-fromJSON}{\code{Localization$fromJSON()}}
\item \href{#method-toJSONString}{\code{Localization$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Localization$fromJSONString()}}
\item \href{#method-clone}{\code{Localization$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Localization$new(
  attributes = NULL,
  email = NULL,
  frame = NULL,
  height = NULL,
  id = NULL,
  media = NULL,
  meta = NULL,
  modified = NULL,
  parent = NULL,
  project = NULL,
  thumbnail_image = NULL,
  u = NULL,
  v = NULL,
  version = NULL,
  width = NULL,
  x = NULL,
  y = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Localization$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Localization$fromJSON(LocalizationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Localization$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Localization$fromJSONString(LocalizationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Localization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
