% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization_spec.R
\docType{class}
\name{LocalizationSpec}
\alias{LocalizationSpec}
\title{LocalizationSpec}
\format{
An \code{R6Class} generator object
}
\description{
LocalizationSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{frame}}{integer}

\item{\code{height}}{numeric \link{optional}}

\item{\code{media_id}}{integer}

\item{\code{modified}}{character \link{optional}}

\item{\code{parent}}{numeric \link{optional}}

\item{\code{type}}{integer}

\item{\code{u}}{numeric \link{optional}}

\item{\code{v}}{numeric \link{optional}}

\item{\code{version}}{integer \link{optional}}

\item{\code{width}}{numeric \link{optional}}

\item{\code{x}}{numeric \link{optional}}

\item{\code{y}}{numeric \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalizationSpec$new()}}
\item \href{#method-toJSON}{\code{LocalizationSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{LocalizationSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{LocalizationSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LocalizationSpec$fromJSONString()}}
\item \href{#method-clone}{\code{LocalizationSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationSpec$new(
  frame,
  media_id,
  type,
  height = NULL,
  modified = NULL,
  parent = NULL,
  u = NULL,
  v = NULL,
  version = NULL,
  width = NULL,
  x = NULL,
  y = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationSpec$fromJSON(LocalizationSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationSpec$fromJSONString(LocalizationSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
