% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_localization_type.R
\docType{class}
\name{LocalizationType}
\alias{LocalizationType}
\title{LocalizationType}
\format{
An \code{R6Class} generator object
}
\description{
LocalizationType Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attribute_types}}{list( \link{AttributeType} ) \link{optional}}

\item{\code{colorMap}}{\link{ColorMap} \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{dtype}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{line_width}}{integer \link{optional}}

\item{\code{media}}{list( integer ) \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{project}}{integer \link{optional}}

\item{\code{visible}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalizationType$new()}}
\item \href{#method-toJSON}{\code{LocalizationType$toJSON()}}
\item \href{#method-fromJSON}{\code{LocalizationType$fromJSON()}}
\item \href{#method-toJSONString}{\code{LocalizationType$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LocalizationType$fromJSONString()}}
\item \href{#method-clone}{\code{LocalizationType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationType$new(
  attribute_types = NULL,
  colorMap = NULL,
  description = NULL,
  dtype = NULL,
  id = NULL,
  line_width = NULL,
  media = NULL,
  name = NULL,
  project = NULL,
  visible = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationType$fromJSON(LocalizationTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationType$fromJSONString(LocalizationTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
