% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization_type_spec.R
\docType{class}
\name{LocalizationTypeSpec}
\alias{LocalizationTypeSpec}
\title{LocalizationTypeSpec}
\format{
An \code{R6Class} generator object
}
\description{
LocalizationTypeSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attribute_types}}{list( \link{AttributeType} ) \link{optional}}

\item{\code{colorMap}}{\link{ColorMap} \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{dtype}}{character}

\item{\code{line_width}}{integer \link{optional}}

\item{\code{media_types}}{list( integer )}

\item{\code{name}}{character}

\item{\code{visible}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalizationTypeSpec$new()}}
\item \href{#method-toJSON}{\code{LocalizationTypeSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{LocalizationTypeSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{LocalizationTypeSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LocalizationTypeSpec$fromJSONString()}}
\item \href{#method-clone}{\code{LocalizationTypeSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeSpec$new(
  dtype,
  media_types,
  name,
  attribute_types = NULL,
  colorMap = NULL,
  description = NULL,
  line_width = NULL,
  visible = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeSpec$fromJSON(LocalizationTypeSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeSpec$fromJSONString(LocalizationTypeSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
