% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization_type_update.R
\docType{class}
\name{LocalizationTypeUpdate}
\alias{LocalizationTypeUpdate}
\title{LocalizationTypeUpdate}
\format{
An \code{R6Class} generator object
}
\description{
LocalizationTypeUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colorMap}}{\link{ColorMap} \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{line_width}}{integer \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{visible}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalizationTypeUpdate$new()}}
\item \href{#method-toJSON}{\code{LocalizationTypeUpdate$toJSON()}}
\item \href{#method-fromJSON}{\code{LocalizationTypeUpdate$fromJSON()}}
\item \href{#method-toJSONString}{\code{LocalizationTypeUpdate$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LocalizationTypeUpdate$fromJSONString()}}
\item \href{#method-clone}{\code{LocalizationTypeUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeUpdate$new(
  colorMap = NULL,
  description = NULL,
  line_width = NULL,
  name = NULL,
  visible = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeUpdate$fromJSON(LocalizationTypeUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeUpdate$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeUpdate$fromJSONString(LocalizationTypeUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationTypeUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
