% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_localization_update.R
\docType{class}
\name{LocalizationUpdate}
\alias{LocalizationUpdate}
\title{LocalizationUpdate}
\format{
An \code{R6Class} generator object
}
\description{
LocalizationUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attributes}}{named list( \link{AnyType} ) \link{optional}}

\item{\code{frame}}{integer \link{optional}}

\item{\code{height}}{numeric \link{optional}}

\item{\code{modified}}{character \link{optional}}

\item{\code{parent}}{numeric \link{optional}}

\item{\code{u}}{numeric \link{optional}}

\item{\code{v}}{numeric \link{optional}}

\item{\code{width}}{numeric \link{optional}}

\item{\code{x}}{numeric \link{optional}}

\item{\code{y}}{numeric \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalizationUpdate$new()}}
\item \href{#method-toJSON}{\code{LocalizationUpdate$toJSON()}}
\item \href{#method-fromJSON}{\code{LocalizationUpdate$fromJSON()}}
\item \href{#method-toJSONString}{\code{LocalizationUpdate$toJSONString()}}
\item \href{#method-fromJSONString}{\code{LocalizationUpdate$fromJSONString()}}
\item \href{#method-clone}{\code{LocalizationUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationUpdate$new(
  attributes = NULL,
  frame = NULL,
  height = NULL,
  modified = NULL,
  parent = NULL,
  u = NULL,
  v = NULL,
  width = NULL,
  x = NULL,
  y = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationUpdate$fromJSON(LocalizationUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationUpdate$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationUpdate$fromJSONString(LocalizationUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalizationUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
