% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\docType{class}
\name{Media}
\alias{Media}
\title{Media}
\format{
An \code{R6Class} generator object
}
\description{
Media Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attributes}}{named list( \link{AnyType} ) \link{optional}}

\item{\code{codec}}{character \link{optional}}

\item{\code{created_by}}{integer \link{optional}}

\item{\code{created_datetime}}{character \link{optional}}

\item{\code{file}}{character \link{optional}}

\item{\code{fps}}{integer \link{optional}}

\item{\code{height}}{integer \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{last_edit_end}}{character \link{optional}}

\item{\code{last_edit_start}}{character \link{optional}}

\item{\code{md5}}{character \link{optional}}

\item{\code{media_files}}{\link{MediaFiles} \link{optional}}

\item{\code{meta}}{integer \link{optional}}

\item{\code{modified_by}}{integer \link{optional}}

\item{\code{modified_datetime}}{character \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{num_frames}}{integer \link{optional}}

\item{\code{original}}{character \link{optional}}

\item{\code{project}}{integer \link{optional}}

\item{\code{segment_info}}{character \link{optional}}

\item{\code{thumbnail}}{character \link{optional}}

\item{\code{thumbnail_gif}}{character \link{optional}}

\item{\code{width}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Media$new()}}
\item \href{#method-toJSON}{\code{Media$toJSON()}}
\item \href{#method-fromJSON}{\code{Media$fromJSON()}}
\item \href{#method-toJSONString}{\code{Media$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Media$fromJSONString()}}
\item \href{#method-clone}{\code{Media$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Media$new(
  attributes = NULL,
  codec = NULL,
  created_by = NULL,
  created_datetime = NULL,
  file = NULL,
  fps = NULL,
  height = NULL,
  id = NULL,
  last_edit_end = NULL,
  last_edit_start = NULL,
  md5 = NULL,
  media_files = NULL,
  meta = NULL,
  modified_by = NULL,
  modified_datetime = NULL,
  name = NULL,
  num_frames = NULL,
  original = NULL,
  project = NULL,
  segment_info = NULL,
  thumbnail = NULL,
  thumbnail_gif = NULL,
  width = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Media$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Media$fromJSON(MediaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Media$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Media$fromJSONString(MediaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Media$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
