% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_media_spec.R
\docType{class}
\name{MediaSpec}
\alias{MediaSpec}
\title{MediaSpec}
\format{
An \code{R6Class} generator object
}
\description{
MediaSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gid}}{character \link{optional}}

\item{\code{md5}}{character}

\item{\code{name}}{character}

\item{\code{progress_name}}{character \link{optional}}

\item{\code{section}}{character}

\item{\code{thumbnail_url}}{character \link{optional}}

\item{\code{type}}{integer}

\item{\code{uid}}{character \link{optional}}

\item{\code{url}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MediaSpec$new()}}
\item \href{#method-toJSON}{\code{MediaSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{MediaSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{MediaSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MediaSpec$fromJSONString()}}
\item \href{#method-clone}{\code{MediaSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaSpec$new(
  md5,
  name,
  section,
  type,
  gid = NULL,
  progress_name = NULL,
  thumbnail_url = NULL,
  uid = NULL,
  url = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaSpec$fromJSON(MediaSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaSpec$fromJSONString(MediaSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
