% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_media_type.R
\docType{class}
\name{MediaType}
\alias{MediaType}
\title{MediaType}
\format{
An \code{R6Class} generator object
}
\description{
MediaType Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{archive_config}}{list( \link{ArchiveConfig} ) \link{optional}}

\item{\code{attribute_types}}{list( \link{AttributeType} ) \link{optional}}

\item{\code{default_volume}}{integer \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{dtype}}{character \link{optional}}

\item{\code{file_format}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{keep_original}}{character \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{project}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MediaType$new()}}
\item \href{#method-toJSON}{\code{MediaType$toJSON()}}
\item \href{#method-fromJSON}{\code{MediaType$fromJSON()}}
\item \href{#method-toJSONString}{\code{MediaType$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MediaType$fromJSONString()}}
\item \href{#method-clone}{\code{MediaType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaType$new(
  archive_config = NULL,
  attribute_types = NULL,
  default_volume = NULL,
  description = "",
  dtype = NULL,
  file_format = NULL,
  id = NULL,
  keep_original = TRUE,
  name = NULL,
  project = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaType$fromJSON(MediaTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaType$fromJSONString(MediaTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
