% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_media_type_spec.R
\docType{class}
\name{MediaTypeSpec}
\alias{MediaTypeSpec}
\title{MediaTypeSpec}
\format{
An \code{R6Class} generator object
}
\description{
MediaTypeSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{archive_config}}{list( \link{ArchiveConfig} ) \link{optional}}

\item{\code{attribute_types}}{list( \link{AttributeType} ) \link{optional}}

\item{\code{default_volume}}{integer \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{dtype}}{character}

\item{\code{file_format}}{character \link{optional}}

\item{\code{keep_original}}{character \link{optional}}

\item{\code{name}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MediaTypeSpec$new()}}
\item \href{#method-toJSON}{\code{MediaTypeSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{MediaTypeSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{MediaTypeSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MediaTypeSpec$fromJSONString()}}
\item \href{#method-clone}{\code{MediaTypeSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaTypeSpec$new(
  dtype,
  name,
  archive_config = NULL,
  attribute_types = NULL,
  default_volume = NULL,
  description = "",
  file_format = NULL,
  keep_original = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaTypeSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaTypeSpec$fromJSON(MediaTypeSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaTypeSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaTypeSpec$fromJSONString(MediaTypeSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaTypeSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
