% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_media_update.R
\docType{class}
\name{MediaUpdate}
\alias{MediaUpdate}
\title{MediaUpdate}
\format{
An \code{R6Class} generator object
}
\description{
MediaUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attributes}}{named list( \link{AnyType} ) \link{optional}}

\item{\code{codec}}{character \link{optional}}

\item{\code{fps}}{numeric \link{optional}}

\item{\code{gid}}{character \link{optional}}

\item{\code{height}}{integer \link{optional}}

\item{\code{last_edit_end}}{character \link{optional}}

\item{\code{last_edit_start}}{character \link{optional}}

\item{\code{media_files}}{\link{MediaFiles} \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{num_frames}}{integer \link{optional}}

\item{\code{thumbnail_gif_url}}{character \link{optional}}

\item{\code{thumbnail_url}}{character \link{optional}}

\item{\code{uid}}{character \link{optional}}

\item{\code{width}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MediaUpdate$new()}}
\item \href{#method-toJSON}{\code{MediaUpdate$toJSON()}}
\item \href{#method-fromJSON}{\code{MediaUpdate$fromJSON()}}
\item \href{#method-toJSONString}{\code{MediaUpdate$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MediaUpdate$fromJSONString()}}
\item \href{#method-clone}{\code{MediaUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaUpdate$new(
  attributes = NULL,
  codec = NULL,
  fps = NULL,
  gid = NULL,
  height = NULL,
  last_edit_end = NULL,
  last_edit_start = NULL,
  media_files = NULL,
  name = NULL,
  num_frames = NULL,
  thumbnail_gif_url = NULL,
  thumbnail_url = NULL,
  uid = NULL,
  width = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaUpdate$fromJSON(MediaUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaUpdate$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaUpdate$fromJSONString(MediaUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MediaUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
