% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_progress_spec.R
\docType{class}
\name{ProgressSpec}
\alias{ProgressSpec}
\title{ProgressSpec}
\format{
An \code{R6Class} generator object
}
\description{
ProgressSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gid}}{character}

\item{\code{job_type}}{character}

\item{\code{media_ids}}{character \link{optional}}

\item{\code{message}}{character}

\item{\code{name}}{character}

\item{\code{progress}}{numeric}

\item{\code{section}}{character \link{optional}}

\item{\code{sections}}{character \link{optional}}

\item{\code{state}}{character}

\item{\code{swid}}{character \link{optional}}

\item{\code{uid}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProgressSpec$new()}}
\item \href{#method-toJSON}{\code{ProgressSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{ProgressSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{ProgressSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ProgressSpec$fromJSONString()}}
\item \href{#method-clone}{\code{ProgressSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressSpec$new(
  gid,
  job_type,
  message,
  name,
  progress,
  state,
  uid,
  media_ids = NULL,
  section = NULL,
  sections = NULL,
  swid = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressSpec$fromJSON(ProgressSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressSpec$fromJSONString(ProgressSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
