% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_type.R
\docType{class}
\name{StateType}
\alias{StateType}
\title{StateType}
\format{
An \code{R6Class} generator object
}
\description{
StateType Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{association}}{character \link{optional}}

\item{\code{attribute_types}}{list( \link{AttributeType} ) \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{dtype}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{interpolation}}{character \link{optional}}

\item{\code{media}}{list( integer ) \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{project}}{integer \link{optional}}

\item{\code{visible}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StateType$new()}}
\item \href{#method-toJSON}{\code{StateType$toJSON()}}
\item \href{#method-fromJSON}{\code{StateType$fromJSON()}}
\item \href{#method-toJSONString}{\code{StateType$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StateType$fromJSONString()}}
\item \href{#method-clone}{\code{StateType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateType$new(
  association = NULL,
  attribute_types = NULL,
  description = NULL,
  dtype = NULL,
  id = NULL,
  interpolation = "latest",
  media = NULL,
  name = NULL,
  project = NULL,
  visible = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateType$fromJSON(StateTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateType$fromJSONString(StateTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
