% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporary_file.R
\docType{class}
\name{TemporaryFile}
\alias{TemporaryFile}
\title{TemporaryFile}
\format{
An \code{R6Class} generator object
}
\description{
TemporaryFile Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{created_datetime}}{character \link{optional}}

\item{\code{eol_datetime}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{lookup}}{character \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{path}}{character \link{optional}}

\item{\code{project}}{integer \link{optional}}

\item{\code{user}}{integer \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TemporaryFile$new()}}
\item \href{#method-toJSON}{\code{TemporaryFile$toJSON()}}
\item \href{#method-fromJSON}{\code{TemporaryFile$fromJSON()}}
\item \href{#method-toJSONString}{\code{TemporaryFile$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TemporaryFile$fromJSONString()}}
\item \href{#method-clone}{\code{TemporaryFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporaryFile$new(
  created_datetime = NULL,
  eol_datetime = NULL,
  id = NULL,
  lookup = NULL,
  name = NULL,
  path = NULL,
  project = NULL,
  user = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporaryFile$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporaryFile$fromJSON(TemporaryFileJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporaryFile$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporaryFile$fromJSONString(TemporaryFileJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemporaryFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
