% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus_base_uploader.R
\docType{class}
\name{TusBaseUploader}
\alias{TusBaseUploader}
\title{TusBaseUploader}
\description{
TusBaseUploader Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TusBaseUploader$new()}}
\item \href{#method-GetHeaders}{\code{TusBaseUploader$GetHeaders()}}
\item \href{#method-GetURLCreationHeaders}{\code{TusBaseUploader$GetURLCreationHeaders()}}
\item \href{#method-ChecksumAlgorithm}{\code{TusBaseUploader$ChecksumAlgorithm()}}
\item \href{#method-ChecksumAlgorithmName}{\code{TusBaseUploader$ChecksumAlgorithmName()}}
\item \href{#method-GetOffset}{\code{TusBaseUploader$GetOffset()}}
\item \href{#method-EncodeMetadata}{\code{TusBaseUploader$EncodeMetadata()}}
\item \href{#method-InitURLAndOffset}{\code{TusBaseUploader$InitURLAndOffset()}}
\item \href{#method-SetURL}{\code{TusBaseUploader$SetURL()}}
\item \href{#method-GetRequestLength}{\code{TusBaseUploader$GetRequestLength()}}
\item \href{#method-GetFileStream}{\code{TusBaseUploader$GetFileStream()}}
\item \href{#method-GetFileSize}{\code{TusBaseUploader$GetFileSize()}}
\item \href{#method-clone}{\code{TusBaseUploader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$new(
  file_path = NULL,
  file_stream = NULL,
  url = NULL,
  client = NULL,
  chunk_size = NULL,
  metadata = NULL,
  retries = 0,
  retry_delay = 30,
  store_url = FALSE,
  url_storage = NULL,
  fingerprinter = NULL,
  upload_checksum = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-GetHeaders}{}}}
\subsection{Method \code{GetHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$GetHeaders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetURLCreationHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-GetURLCreationHeaders}{}}}
\subsection{Method \code{GetURLCreationHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$GetURLCreationHeaders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChecksumAlgorithm"></a>}}
\if{latex}{\out{\hypertarget{method-ChecksumAlgorithm}{}}}
\subsection{Method \code{ChecksumAlgorithm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$ChecksumAlgorithm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChecksumAlgorithmName"></a>}}
\if{latex}{\out{\hypertarget{method-ChecksumAlgorithmName}{}}}
\subsection{Method \code{ChecksumAlgorithmName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$ChecksumAlgorithmName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetOffset"></a>}}
\if{latex}{\out{\hypertarget{method-GetOffset}{}}}
\subsection{Method \code{GetOffset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$GetOffset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncodeMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-EncodeMetadata}{}}}
\subsection{Method \code{EncodeMetadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$EncodeMetadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InitURLAndOffset"></a>}}
\if{latex}{\out{\hypertarget{method-InitURLAndOffset}{}}}
\subsection{Method \code{InitURLAndOffset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$InitURLAndOffset(url = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetURL"></a>}}
\if{latex}{\out{\hypertarget{method-SetURL}{}}}
\subsection{Method \code{SetURL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$SetURL(url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRequestLength"></a>}}
\if{latex}{\out{\hypertarget{method-GetRequestLength}{}}}
\subsection{Method \code{GetRequestLength()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$GetRequestLength()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileStream"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileStream}{}}}
\subsection{Method \code{GetFileStream()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$GetFileStream()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFileSize"></a>}}
\if{latex}{\out{\hypertarget{method-GetFileSize}{}}}
\subsection{Method \code{GetFileSize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$GetFileSize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusBaseUploader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
