% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus_client.R
\docType{class}
\name{TusClient}
\alias{TusClient}
\title{TusClient}
\description{
TusClient Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TusClient$new()}}
\item \href{#method-SetHeaders}{\code{TusClient$SetHeaders()}}
\item \href{#method-Uploader}{\code{TusClient$Uploader()}}
\item \href{#method-clone}{\code{TusClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$new(url = NULL, headers = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SetHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-SetHeaders}{}}}
\subsection{Method \code{SetHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$SetHeaders(headers = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uploader"></a>}}
\if{latex}{\out{\hypertarget{method-Uploader}{}}}
\subsection{Method \code{Uploader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$Uploader(
  file_path = NULL,
  chunk_size = NULL,
  retries = 10,
  retry_delay = 15
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
