% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tus_uploader.R
\docType{class}
\name{TusUploader}
\alias{TusUploader}
\title{TusUploader}
\description{
TusUploader Class
}
\section{Super class}{
\code{\link[tator:TusBaseUploader]{tator::TusBaseUploader}} -> \code{TusUploader}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TusUploader$new()}}
\item \href{#method-UploadChunk}{\code{TusUploader$UploadChunk()}}
\item \href{#method-CreateURL}{\code{TusUploader$CreateURL()}}
\item \href{#method-DoRequest}{\code{TusUploader$DoRequest()}}
\item \href{#method-Retry}{\code{TusUploader$Retry()}}
\item \href{#method-clone}{\code{TusUploader$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="ChecksumAlgorithm">}\href{../../tator/html/TusBaseUploader.html#method-ChecksumAlgorithm}{\code{tator::TusBaseUploader$ChecksumAlgorithm()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="ChecksumAlgorithmName">}\href{../../tator/html/TusBaseUploader.html#method-ChecksumAlgorithmName}{\code{tator::TusBaseUploader$ChecksumAlgorithmName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="EncodeMetadata">}\href{../../tator/html/TusBaseUploader.html#method-EncodeMetadata}{\code{tator::TusBaseUploader$EncodeMetadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="GetFileSize">}\href{../../tator/html/TusBaseUploader.html#method-GetFileSize}{\code{tator::TusBaseUploader$GetFileSize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="GetFileStream">}\href{../../tator/html/TusBaseUploader.html#method-GetFileStream}{\code{tator::TusBaseUploader$GetFileStream()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="GetHeaders">}\href{../../tator/html/TusBaseUploader.html#method-GetHeaders}{\code{tator::TusBaseUploader$GetHeaders()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="GetOffset">}\href{../../tator/html/TusBaseUploader.html#method-GetOffset}{\code{tator::TusBaseUploader$GetOffset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="GetRequestLength">}\href{../../tator/html/TusBaseUploader.html#method-GetRequestLength}{\code{tator::TusBaseUploader$GetRequestLength()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="GetURLCreationHeaders">}\href{../../tator/html/TusBaseUploader.html#method-GetURLCreationHeaders}{\code{tator::TusBaseUploader$GetURLCreationHeaders()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="InitURLAndOffset">}\href{../../tator/html/TusBaseUploader.html#method-InitURLAndOffset}{\code{tator::TusBaseUploader$InitURLAndOffset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tator" data-topic="TusBaseUploader" data-id="SetURL">}\href{../../tator/html/TusBaseUploader.html#method-SetURL}{\code{tator::TusBaseUploader$SetURL()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$new(client = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UploadChunk"></a>}}
\if{latex}{\out{\hypertarget{method-UploadChunk}{}}}
\subsection{Method \code{UploadChunk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$UploadChunk()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateURL"></a>}}
\if{latex}{\out{\hypertarget{method-CreateURL}{}}}
\subsection{Method \code{CreateURL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$CreateURL()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DoRequest"></a>}}
\if{latex}{\out{\hypertarget{method-DoRequest}{}}}
\subsection{Method \code{DoRequest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$DoRequest()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Retry"></a>}}
\if{latex}{\out{\hypertarget{method-Retry}{}}}
\subsection{Method \code{Retry()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$Retry()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TusUploader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
