% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\docType{class}
\name{Version}
\alias{Version}
\title{Version}
\format{
An \code{R6Class} generator object
}
\description{
Version Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bases}}{list( integer ) \link{optional}}

\item{\code{created_by}}{character \link{optional}}

\item{\code{created_datetime}}{character \link{optional}}

\item{\code{description}}{character \link{optional}}

\item{\code{id}}{integer \link{optional}}

\item{\code{modified_by}}{character \link{optional}}

\item{\code{modified_datetime}}{character \link{optional}}

\item{\code{name}}{character \link{optional}}

\item{\code{num_created}}{integer \link{optional}}

\item{\code{num_modified}}{integer \link{optional}}

\item{\code{number}}{integer \link{optional}}

\item{\code{project}}{integer \link{optional}}

\item{\code{show_empty}}{character \link{optional}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Version$new()}}
\item \href{#method-toJSON}{\code{Version$toJSON()}}
\item \href{#method-fromJSON}{\code{Version$fromJSON()}}
\item \href{#method-toJSONString}{\code{Version$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Version$fromJSONString()}}
\item \href{#method-clone}{\code{Version$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Version$new(
  bases = NULL,
  created_by = NULL,
  created_datetime = NULL,
  description = "",
  id = NULL,
  modified_by = NULL,
  modified_datetime = NULL,
  name = NULL,
  num_created = NULL,
  num_modified = NULL,
  number = NULL,
  project = NULL,
  show_empty = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Version$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Version$fromJSON(VersionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Version$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Version$fromJSONString(VersionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Version$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
