% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_spec.R
\docType{class}
\name{VideoSpec}
\alias{VideoSpec}
\title{VideoSpec}
\format{
An \code{R6Class} generator object
}
\description{
VideoSpec Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{codec}}{character}

\item{\code{fps}}{numeric}

\item{\code{gid}}{character}

\item{\code{height}}{integer}

\item{\code{md5}}{character}

\item{\code{media_files}}{\link{MediaFiles}}

\item{\code{name}}{character}

\item{\code{num_frames}}{integer}

\item{\code{progressName}}{character \link{optional}}

\item{\code{section}}{character}

\item{\code{thumbnail_gif_url}}{character}

\item{\code{thumbnail_url}}{character}

\item{\code{type}}{integer}

\item{\code{uid}}{character}

\item{\code{width}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VideoSpec$new()}}
\item \href{#method-toJSON}{\code{VideoSpec$toJSON()}}
\item \href{#method-fromJSON}{\code{VideoSpec$fromJSON()}}
\item \href{#method-toJSONString}{\code{VideoSpec$toJSONString()}}
\item \href{#method-fromJSONString}{\code{VideoSpec$fromJSONString()}}
\item \href{#method-clone}{\code{VideoSpec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoSpec$new(
  codec,
  fps,
  gid,
  height,
  md5,
  media_files,
  name,
  num_frames,
  section,
  thumbnail_gif_url,
  thumbnail_url,
  type,
  uid,
  width,
  progressName = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoSpec$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoSpec$fromJSON(VideoSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoSpec$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoSpec$fromJSONString(VideoSpecJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VideoSpec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
